Fit a graph-regularized linear regression model using edge-based regularization.

Fit a graph-regularized linear regression model using edge-penalization. The coefficients are computed using graph-prior knowledge in the form of one/two affinity matrices. Graph-regularization is an extension to previously introduced regularization techniques, such as the LASSO. See the vignette for details on the objective function of the model: vignette("edgenet", package="netReg")

edgenet(X, Y, G.X = NULL, G.Y = NULL, lambda = 1, psigx = 1,
  psigy = 1, thresh = 1e-05, maxit = 1e+05, family = gaussian)

Arguments

X
input matrix, of dimension (n x p) where n is the number of observations and p is the number of covariables. Each row is an observation vector.

Y
output matrix, of dimension (n x q) where n is the number of observations and q is the number of response variables. Each row is an observation vector.

G.X
non-negativ affinity matrix for n, of dimensions (p x p) where p is the number of covariables X

G.Y
non-negativ affinity matrix for n, of dimensions (q x q) where q is the number of responses Y

lambda
numerical shrinkage parameter for LASSO.

psigx
numerical shrinkage parameter for graph-regularization of G.X

psigy
numerical shrinkage parameter for graph-regularization of G.Y

thresh
numerical threshold for optimizer

maxit
maximum number of iterations for optimizer (integer)

family
family of response, e.g. gaussian or binomial

Value

An object of class edgenet

coefficients

the estimated (p x q)-dimensional coefficient matrix B.hat

intercept

the estimated (q x 1)-dimensional vector of intercepts

call

the call that produced the object

family

the family of the response

References

Dirmeier, Simon and Fuchs, Christiane and Mueller, Nikola S and Theis, Fabian J (2018), netReg: Network-regularized linear models for biological association studies.
Bioinformatics

Friedman J., Hastie T., Hoefling H. and Tibshirani R. (2007), Pathwise coordinate optimization.
The Annals of Applied Statistics

Friedman J., Hastie T. and Tibshirani R. (2010), Regularization Paths for Generalized Linear Models via Coordinate Descent.
Journal of Statistical Software

Fu W. J. (1998), Penalized Regression: The Bridge Versus the Lasso.
Journal of Computational and Graphical Statistics

Cheng W. and Wang W. (2014), Graph-regularized dual Lasso for robust eQTL mapping.
Bioinformatics

Examples

X <- matrix(rnorm(100*10), 100, 10)
b <- rnorm(10)
G.X <- matrix(rpois(100,1), 10)
G.X <- t(G.X) + G.X
diag(G.X) <- 0

# fit a Gaussian model
Y <- X %*% b + rnorm(100)
fit <- edgenet(X=X, Y=Y, G.X=G.X, family=gaussian)