The netReg package

netReg is a package for generalized linear regression that includes prior graphs in the models objective function.

Details

netReg uses Armadillo and TensorFlow for fast matrix computations and optimization.

References

Dirmeier, Simon and Fuchs, Christiane and Mueller, Nikola S and Theis, Fabian J (2018), netReg: Network-regularized linear models for biological association studies.
Bioinformatics

Abadi, Martín et al. (2016), Tensorflow: A system for large-scale machine learning.
12th USENIX Symposium on Operating Systems Design and Implementation (OSDI 16)

Powell M.J.D. (2009), The BOBYQA algorithm for bound constrained optimization without derivatives.
http://www.damtp.cam.ac.uk/user/na/NA_papers/NA2009_06.pdf Eddelbuettel, Dirk and Sanderson, Conrad (2014), RcppArmadillo: Accelerating R with high-performance C++ linear algebra. Computational Statistics & Data Analysis