Find the optimal shrinkage parameters for edgenet

Finds the optimal shrinkage parameters using cross-validation for edgenet. We use the BOBYQA algorithm to minimize the sum of squared residuals objective function.

cv.edgenet(X, Y, G.X = NULL, G.Y = NULL, thresh = 1e-05, maxit = 1e+05,
  family = c("gaussian"), epsilon = 0.001, approx.maxit = 10000,
  nfolds = 10, ...)

Arguments

X
input matrix, of dimension (n x p) where n is the number of observations and p is the number of covariables. Each row is an observation vector.

Y
output matrix, of dimension (n x q) where n is the number of observations and q is the number of response variables Each row is an observation vector.

G.X
non-negativ affinity matrix for n, of dimensions (p x p) where p is the number of covariables X

G.Y
non-negativ affinity matrix for n, of dimensions (q x q) where q is the number of covariables Y

thresh
threshold for coordinate descent

maxit
maximum number of iterations

family
family of response, e.g. gaussian

epsilon
the threshold criterion for BOBYQA to stop. Usually 1e-3 is a good choice.

approx.maxit
the maximum number of iterations for BOBYQA (if choosen). Usually 1e4 is a good choice.

nfolds
the number of folds to be used - default is 10 (minimum 3, maximum nrow(X)).

...
additional parameters

Value

An object of class cv.edgenet

call

the call that produced the object

lambda

the estimated (p x q)-dimensional coefficient matrix B.hat

psigx

the estimated (q x 1)-dimensional vector of intercepts

psigy

the estimated (q x 1)-dimensional vector of intercepts

References

Friedman J., Hastie T., Hoefling H. and Tibshirani R. (2007), Pathwise coordinate optimization. The Annals of Applied Statistics Friedman J., Hastie T. and Tibshirani R. (2010), Regularization Paths for Generalized Linear Models via Coordinate Descent. Journal of Statistical Software Fu W. J. (1998), Penalized Regression: The Bridge Versus the Lasso. Journal of Computational and Graphical Statistics Cheng W. and Wang W. (2014), Graph-regularized dual Lasso for robust eQTL mapping. Bioinformatics Powell M.J.D. (2009), The BOBYQA algorithm for bound constrained optimization without derivatives. http://www.damtp.cam.ac.uk/user/na/NA_papers/NA2009_06.pdf

Examples

X <- matrix(rnorm(100*10), 100, 10)
b <- rnorm(10)
G.X <- matrix(rpois(10*10,1),10)
G.X <- t(G.X) + G.X
diag(G.X) <- 0

# fit a Gaussian model
Y <- X%*%b + rnorm(100)
cv.edge <- cv.edgenet(X=X, Y=Y, G.X=G.X, family="gaussian")