netReg

netReg is a package for generalized linear regression that includes prior graphs in the models objective function.

Details

netReg uses Armadillo, OpenBLAS, BLAS and LAPACK for fast matrix computations and Dlib for constrained derivate-free optimization.

References

Friedman J., Hastie T., Hoefling H. and Tibshirani R. (2007), Pathwise coordinate optimization. The Annals of Applied Statistics Friedman J., Hastie T. and Tibshirani R. (2010), Regularization Paths for Generalized Linear Models via Coordinate Descent. Journal of Statistical Software Fu W. J. (1998), Penalized Regression: The Bridge Versus the Lasso. Journal of Computational and Graphical Statistics Cheng W. and Wang W. (2014), Graph-regularized dual Lasso for robust eQTL mapping. Bioinformatics Powell M.J.D. (2009), The BOBYQA algorithm for bound constrained optimization without derivatives. http://www.damtp.cam.ac.uk/user/na/NA_papers/NA2009_06.pdf